+++ lib/src/methods/methods.dart
@@ -3,3 +3,4 @@ export 'user_info/user_info_methods.dart';
 export 'trade/trade_methods.dart';
 export 'updates/updates_methods.dart';
 export 'other/other_methods.dart';
+export 'nft/nft_methods.dart';